+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## vNext
 * WebApps/Functions: Don't turn on Logging Extension for Linux App Service.
+* WebApps/Functions: Fix .NET 5/6 on Linux deployments.
 
 ## 1.6.25
 * CosmosDb: Add support for serverless capacity mode.