+++ RELEASE_NOTES.md
@@ -1,9 +1,11 @@
 Release Notes
 =============
 ## vNext
+* WebApp: Support custom port for docker container with `docker_port`
+
+## 1.6.26
 * WebApps/Functions: Don't turn on Logging Extension for Linux App Service.
 * WebApps/Functions: Fix .NET 5/6 on Linux deployments.
-* WebApp: Support custom port for docker container with `docker_port`
 
 ## 1.6.25
 * CosmosDb: Add support for serverless capacity mode.